mqtt: widen VBI multiplier to avoid silent truncation past 2 bytes
mqtt_decode_var_byte_int() declared multiplier as uint8_t and updated it
as `multiplier *= 128`. The first iteration produces 128 (still uint8_t)
but the second wraps to 0, so every contribution from the third and
fourth bytes of a Variable Byte Integer is multiplied by 0 and silently
dropped. The decoder then returns a small value with all higher bytes
discarded.

This affects remaining_length and v5 property-length parsing for any
value above 16,383, which is most PUBLISH payloads larger than 16 KB.

Use uint32_t for multiplier so all four spec-allowed VBI bytes
contribute correctly within the existing 16-bit dest range.

+++ os/net/app-layer/mqtt/mqtt.c
@@ -382,7 +382,7 @@ mqtt_decode_var_byte_int(const uint8_t *input_data_ptr,
 {
   uint8_t read_bytes = 0;
   uint8_t byte_in;
-  uint8_t multiplier = 1;
+  uint32_t multiplier = 1;
   uint32_t input_pos_0 = 0;
 
   if(input_pos == NULL) {
